requirement.txt
pyspark 
mysql-connector-python 
pandas